refactor: split types/geospatial.go into config/bounds and WKB parsing
Extract WKB geometry parsing functions into geospatial_wkb.go; split
matching tests into geospatial_wkb_test.go.
